Add files via upload
In this script, we merge our dataframe with outcome variables (change in mme before and after August 2013) with the County Health Rankings database from University of Wisconsin Population Health Institute (linked here: https://www.countyhealthrankings.org/health-data/new-york/data-and-resources). As the number of features in comparison to observations is large, we apply recursive feature elimination to select for the most predictive variables and avoid overfitting our model. As there is an imbalance in outcomes within our dataset (41 counties decreased vs 9 increased), we also have applied synthetic minority oversampling technique when training the model. We have also generated an ROC and classification matrix to help evaluate performance of the model in this script.